+++ test/option/parser_test.rb
@@ -46,7 +46,18 @@ class ParserTest < Test::Unit::TestCase
     assert_nil(options[Maze::Option::UDID])
 
     # Logging options
-    assert_true(options[Maze::Option::LOG_REQUESTS])
+    assert_true(options[Maze::Option::FILE_LOG])
+    assert_false(options[Maze::Option::LOG_REQUESTS])
+  end
+
+  def test_buildkite_default_values
+    ENV['BUILDKITE'] = "true"
+
+    args = %w[]
+    options = Maze::Option::Parser.parse args
+
+    # Logging options
+    assert_false(options[Maze::Option::LOG_REQUESTS])
   end
 
   def test_overwritten_values
@@ -69,7 +80,8 @@ class ParserTest < Test::Unit::TestCase
       --no-start-appium
       --apple-team-id=ARG_APPLE_TEAM_ID
       --udid=ARG_UDID
-      --no-log-requests
+      --log-requests
+      --no-file-log
     ]
     options = Maze::Option::Parser.parse args
 
@@ -97,8 +109,9 @@ class ParserTest < Test::Unit::TestCase
     assert_equal('ARG_APPLE_TEAM_ID', options[Maze::Option::APPLE_TEAM_ID])
     assert_equal('ARG_UDID', options[Maze::Option::UDID])
 
-    # Logger options
-    assert_false(options[Maze::Option::LOG_REQUESTS])
+    # Logging options
+    assert_false(options[Maze::Option::FILE_LOG])
+    assert_true(options[Maze::Option::LOG_REQUESTS])
   end
 
   def test_short_flags