+++ test/option/processor_test.rb
@@ -9,6 +9,7 @@ require_relative '../../lib/maze/helper'
 # Tests the options parser and processor together (using only valid options and with no validator).
 class ProcessorTest < Test::Unit::TestCase
   def setup
+    ENV.delete('BUILDKITE')
     ENV.delete('MAZE_BS_LOCAL')
     ENV.delete('MAZE_SL_LOCAL')
     ENV.delete('BROWSER_STACK_USERNAME')
@@ -67,12 +68,13 @@ class ProcessorTest < Test::Unit::TestCase
   end
 
   def test_logger_options
-    args = %w[--no-log-requests --always-log]
+    args = %w[--no-file-log --log-requests --always-log]
     options = Maze::Option::Parser.parse args
     config = Maze::Configuration.new
     Maze::Option::Processor.populate config, options
 
-    assert_false config.log_requests
+    assert_false config.file_log
+    assert_true config.log_requests
     assert_true config.always_log
   end
 
@@ -90,7 +92,8 @@ class ProcessorTest < Test::Unit::TestCase
     assert_false config.resilient
     assert_nil config.capabilities
 
-    assert_true config.log_requests
+    assert_true config.file_log
+    assert_false config.log_requests
     assert_false config.always_log
   end
 end